mysql novo/pp/requirements.txt
altgraph==0.17
antiorm==1.2.1
asgiref==3.3.4
astroid==2.5.2
atomicwrites==1.4.0
attrs==20.3.0
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cs50==6.0.4
cx-Freeze==6.5.3
db==0.1.1
db-sqlite3==0.0.1
Django==3.2
dnspython==2.1.0
Flask==1.1.2
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.0.0
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector-python==8.0.24
mysqlclient==2.0.3
packaging==20.9
pefile==2019.4.18
pluggy==0.13.1
protobuf==3.15.8
py==1.10.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyinstaller-hooks-contrib==2021.1
pylint==2.7.3
pymongo==3.11.3
PyMsgBox==1.0.9
pyngrok==5.0.5
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytest==6.2.3
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sql==0.4.0
SQLAlchemy==1.4.11
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
